[tests] fix varray util func imports
