+++ src/pymortests/algorithms/basic.py
@@ -13,8 +13,8 @@ from pymor.operators.constructions import induced_norm
 from pymor.operators.numpy import NumpyMatrixOperator
 from pymor.vectorarrays.numpy import NumpyVectorSpace
 from pymortests.fixtures.operator import operator_with_arrays_and_products
-from pymortests.strategies import valid_inds, valid_inds_of_same_length
-from pymortests.vectorarray import invalid_ind_pairs, indexed
+from pymortests.strategies import valid_inds, valid_inds_of_same_length, invalid_ind_pairs
+from pymortests.vectorarray import indexed
 import pymortests.strategies as pyst
